--- packages/tpetra/core/src/Tpetra_Details_normImpl_decl.hpp
@@ -1,87 +1,0 @@
-// @HEADER
-// ***********************************************************************
-//
-//          Tpetra: Templated Linear Algebra Services Package
-//                 Copyright (2008) Sandia Corporation
-//
-// Under the terms of Contract DE-AC04-94AL85000 with Sandia Corporation,
-// the U.S. Government retains certain rights in this software.
-//
-// Redistribution and use in source and binary forms, with or without
-// modification, are permitted provided that the following conditions are
-// met:
-//
-// 1. Redistributions of source code must retain the above copyright
-// notice, this list of conditions and the following disclaimer.
-//
-// 2. Redistributions in binary form must reproduce the above copyright
-// notice, this list of conditions and the following disclaimer in the
-// documentation and/or other materials provided with the distribution.
-//
-// 3. Neither the name of the Corporation nor the names of the
-// contributors may be used to endorse or promote products derived from
-// this software without specific prior written permission.
-//
-// THIS SOFTWARE IS PROVIDED BY SANDIA CORPORATION "AS IS" AND ANY
-// EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
-// PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SANDIA CORPORATION OR THE
-// CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
-// EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
-// PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
-// PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
-// LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
-// NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-// SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-//
-// Questions? Contact Michael A. Heroux (maherou@sandia.gov)
-//
-// ************************************************************************
-// @HEADER
-
-#ifndef TPETRA_DETAILS_NORMIMPL_DECL_HPP
-#define TPETRA_DETAILS_NORMIMPL_DECL_HPP
-
-/// \file Tpetra_Details_normImpl_def.hpp
-/// \brief Definition of the Tpetra::Details::normImpl function
-
-#include "TpetraCore_config.h"
-#include "Kokkos_Core.hpp"
-
-#ifndef DOXYGEN_SHOULD_SKIP_THIS
-namespace Teuchos {
-  template<class T>
-  class ArrayView; // forward declaration
-  template<class OrdinalType>
-  class Comm; // forward declaration
-}
-#endif // DOXYGEN_SHOULD_SKIP_THIS
-
-namespace Tpetra {
-namespace Details {
-
-//! Input argument for normImpl() (which see).
-enum EWhichNorm {
-  NORM_ONE, //<! Use the one-norm
-  NORM_TWO, //<! Use the two-norm
-  NORM_INF  //<! Use the infinity-norm
-};
-
-//! Implementation of MultiVector norms.
-template <class ValueType,
-          class ArrayLayout,
-          class DeviceType,
-          class MagnitudeType>
-void
-normImpl (MagnitudeType norms[],
-          const Kokkos::View<const ValueType**, ArrayLayout, DeviceType>& X,
-          const EWhichNorm whichNorm,
-          const Teuchos::ArrayView<const size_t>& whichVecs,
-          const bool isConstantStride,
-          const bool isDistributed,
-          const Teuchos::Comm<int>* comm);
-
-} // namespace Details
-} // namespace Tpetra
-
-#endif // TPETRA_DETAILS_NORMIMPL_DECL_HPP
